rename(dto, domain): dto, domain 클래스 추가 dto - input, output 구분
