+++ src/test/java/lotto/userMoneyDTOTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.UserMoney;
+import lotto.dto.input.UserMoneyDTO;
 import lotto.util.Parser;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
